refactor: split docs extractor visitor
